lint
SCCS-vsn: sys/vax/uba/dhu.c 4.9
SCCS-vsn: sys/vax/uba/tm.c 6.9
SCCS-vsn: sys/vax/uba/tmscp.c 5.2
SCCS-vsn: sys/vax/uba/ts.c 6.9
SCCS-vsn: sys/vax/uba/uda.c 6.18
SCCS-vsn: sys/vax/uba/ut.c 6.7
SCCS-vsn: sys/vax/uba/vp.c 6.7
SCCS-vsn: sys/vax/if/if_vv.c 6.16

+++ usr/src/sys/vax/uba/ut.c
@@ -3,7 +3,7 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)ut.c	6.6 (Berkeley) %G%
+ *	@(#)ut.c	6.7 (Berkeley) %G%
  */
 
 #include "tj.h"
@@ -586,7 +586,7 @@ ignoreerr:
 
 	case SCOM:		/* motion commands update current position */
 		if (bp == &cutbuf[UTUNIT(bp->b_dev)])
-		switch (bp->b_command) {
+		switch ((int)bp->b_command) {
 
 		case UT_SFORW:
 			sc->sc_blkno -= bp->b_repcnt;
